Move server components to src dir
